+++ pytest_xvfb.py
@@ -52,8 +52,10 @@ class Xvfb(object):
             self._virtual_display.wait()  # collect outputs
             stdout = self._virtual_display.stdout
             stderr = self._virtual_display.stderr
-            raise XvfbExitedError("Xvfb exited with exit code {0} stdout={1!r} stderr={2!r}".format(
-                ret, stdout, stderr))
+            raise XvfbExitedError("Xvfb exited with exit code {0}\nXvfb stdout:\n    {1}\nXvfb stderr:\n    {2}".format(
+                ret,
+                '\n    '.join(stdout.splitlines()),
+                '\n    '.join(stderr.splitlines())))
 
     def stop(self):
         self._virtual_display.stop()